major changes
Added contact.html for a thankyou page after contact is submitted. Added Embeded Google Form to send notification for email. Changed the Subject field and added a Name field in contact form.  Added the same code to the merchandise page for the form. Added a Back to home button in the code. Commented out the current Subject code in case I can get it to work in the future.